better componentize sidebar and mobile actions
